Fix reward scaler when run on varied episode lengths
When calling `fit` with a reward scaler on a dataset with varied episode lengths,
the following error would be thrown in the `fit_with_trajectory_slicer` method:

```
ValueError: setting an array element with a sequence. The requested array has an
inhomogeneous shape after 1 dimensions.
```

This commit fixes the issue by flattening the rewards before calculating the mean and std.